Creating a new branch register_middlwares_gates

+++ server/app/Http/Controllers/api/EnseignantController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\api;
+use App\Events\storeuser;
 use App\Http\Controllers\AuthController;
 use App\Http\Resources\EnseignantResource;
 
@@ -14,6 +15,7 @@ use App\Models\Intervention;
 use App\Models\Paiements;
 use App\Traits\HttpResponses;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EnseignantController extends Controller
 {
@@ -21,7 +23,8 @@ class EnseignantController extends Controller
      // A class that handles the success and error messages
      use HttpResponses;
 //=========================================================== The access is retricted for:AdminUae|President ==============================================
-        
+
+    
  
      /**
      * Index() it's a methode that serve to display all the profs with there etablissement and grade.
@@ -32,9 +35,10 @@ class EnseignantController extends Controller
        public function index()
 
           {   
-                       
-                 return EnseignantResource::collection(Enseignant::with('etablissement','grade')->latest()->paginate(10)); 
-                                
+            if (Gate::allows('check_role', [0])) {          
+           return EnseignantResource::collection(Enseignant::with('etablissement','grade')->latest()->paginate(10)); 
+           }
+            return $this->error('','ACCES INTERDIT ',403);                    
           }
 
 //============================================================== The access is retricted for:AdminUEtab ========================================================          
@@ -51,25 +55,36 @@ class EnseignantController extends Controller
 
           {  
 
+              
+              if (Gate::allows('check_role', [1])) { 
                  $enseignant=new Enseignant();
+
                  $grade_id=$enseignant->IdGrade($request['Grade']);                                  //IdGrade() it's a method that return the id of the grade 
                  $etablissement_id=auth()->user()->administrateur->etablissement_id;                //the security developper should approuve it
                  $enseignant->PPR = $request['PPR'];
                  $enseignant->nom = $request['nom'];
                  $enseignant->prenom = $request['prenom'];
                  $enseignant->date_naissance = $request['DateNaissance'];
-                 $enseignant->etablissement_id = $etablissement_id;
-                 $enseignant->grade_id =$grade_id; 
-                  $enseignant->email_perso=$request['email_perso'];                                 
+                 $enseignant->etablissement_id = auth()->user()->administrateur->etablissement_id  ;
+                 $enseignant->grade_id =$grade_id;  
+                                                 
+                 $enseignant->email_perso=$request['email_perso'];
                  $enseignant->save();
-                 return $this->succes("","added successfully");
-        
+                
+                 $id=event (new storeuser($request->input('email_perso'),4,$request->input('nom'),$request->input('prenom')));
+                 $enseignant->user_id = $id[0];
+                 $enseignant->save();
+
+
+                 return $this->succes("","added successfully");}
+                 return $this->error('','ACCES INTERDIT ',403);
+
                  // return new EnseignantResource(Enseignant::create($request->all()));it's another method but we should know the etab_id 
 
           }
 //======================================================= The access is retricted for:AdminUEtab|AdminUae|DireteurEtab|President ==================================================
      
-    
+    //the access to this method is restricted for 0,3 and also 1 and 2
      /**
      * Show() this method serve to display a specified teacher with the name and the city of his etablissement and also his  grade  .
      * @param  int  $id it's IDEnseignant!!!!!!
@@ -78,8 +93,12 @@ class EnseignantController extends Controller
    
        public function show($id)
 
-          {                                    
-                 return new EnseignantResource(Enseignant::with('etablissement','grade')->FindOrFail($id));                    
+          {      if (Gate::allows('check_role', [0,3]) || Gate::allows('direct_ens',$id) || Gate::allows('admin_ens',$id) ) {  
+
+                 return new EnseignantResource(Enseignant::with('etablissement','grade')->FindOrFail($id));
+                }
+                 
+                return $this->error('','ACCES INTERDIT ',403);                   
                  
           }
         
@@ -96,7 +115,7 @@ class EnseignantController extends Controller
 
        public function update( UpdateEnseignantRequest $request, $id)
 
-          {               
+          {       if ( Gate::allows('admin_ens',$id)) {             
                       
                  $enseignant=Enseignant::find($id);
                  $grade_id=$enseignant->IdGrade($request['Grade']);  
@@ -108,6 +127,8 @@ class EnseignantController extends Controller
                 
                  $enseignant->save();
                  return $this->succes("","updated successfully");
+                }
+                 return $this->error('','ACCES INTERDIT ',403);  
                               
           }
 //======================================= The access is retricted for:AdminEtab ==============================================
@@ -121,12 +142,15 @@ class EnseignantController extends Controller
 
        public function destroy($id)
 
-          {                
+
+          {       
+              if ( Gate::allows('admin_ens',$id) ) {           
                  $ens= Enseignant::FindOrfail($id);
-                 if($ens->image)              
-                 {unlink(public_path('uploads').'/'.$ens->image);};                                         //destroy the appropriate image .       
+                 if($ens->image)
+                 unlink(public_path('uploads').'/'.$ens->image);                                         //destroy the appropriate image .       
                  $ens->delete();              
-                 return $this->succes("","enseignant deleted successfully");  
+                 return $this->succes("","enseignant deleted successfully"); }
+                 return $this->error('','ACCES INTERDIT ',403);  
           }
 
 //=============================================== The access is retricted for:Enseignant ===========================================================================                
@@ -138,10 +162,11 @@ class EnseignantController extends Controller
      */
 
 
+
        public function ShowMyInterventions()
 
 
-          {     
+          {   if ( auth()->user()->role==4 ){  
               $id=auth()->user()->enseignant->id;
               
               if(Intervention::where('enseignant_id',$id)->exists())
@@ -154,6 +179,8 @@ class EnseignantController extends Controller
               {
                      return $this->error("","Pas d'interventions pour le moment",404); 
               }
+          }
+           //add Return access denied
                
 
              
@@ -172,7 +199,7 @@ class EnseignantController extends Controller
            
        public function ShowMyPayments()
 
-          {      
+          {      if(auth()->user()->role==4){
                  $id=auth()->user()->enseignant->id;
 
                  if(Paiements::where('enseignant_id',$id)->exists())
@@ -189,6 +216,7 @@ class EnseignantController extends Controller
                      return $this->error("",'Pas de payements pour le moment',404);
                  }
           }
+          }
    
 
 
@@ -204,8 +232,8 @@ class EnseignantController extends Controller
        public function UploadMyImage( Request $request)
        
           {
-                 $id=auth()->user()->enseignant->id;
-
+              $id=auth()->user()->enseignant->id;
+              if (Gate::allows('ens_check_id', $id)) {
                  $request->validate([ 'image'=>'required|max:1024|mimes:png,jpg,png' ]);
 
                  $enseignant=Enseignant::where('id',$id)->first();
@@ -226,6 +254,9 @@ class EnseignantController extends Controller
                  {
                  return $this->succes("","image uploaded successfully");    
                  }
+              }
+              return $this->error('','ACCES INTERDIT ',403); 
+
           }
 //==================================== The access is retricted for:Enseignant   ============================================================                
            
@@ -239,8 +270,11 @@ class EnseignantController extends Controller
        public function ShowMyProfil()
 
           {
-                 $id=auth()->user()->enseignant->id;
-                 return new EnseignantResource(Enseignant::where('id',$id)->with('etablissement','grade')->first());
+              $id=auth()->user()->enseignant->id;
+              if (Gate::allows('ens_check_id', $id)) {
+                     return new EnseignantResource(Enseignant::where('user_id',$id)->with('etablissement','grade')->first());
+              }
+              return $this->error('','ACCES INTERDIT ',403);   
           }
 
 //======================================= The access is retricted for:Enseignant ======================================= 
@@ -250,10 +284,10 @@ class EnseignantController extends Controller
      * MyHours this method serve count The hours worked for a specified enseignant .
      */
        public function MyHours ()
-
           {
+              if (Gate::allows('check_role', [0])){
               $id=auth()->user()->enseignant->id;
-              
+               
               if(Intervention::where('enseignant_id',$id)->exists())
               {
                  $enseignant=new Enseignant();
@@ -262,6 +296,7 @@ class EnseignantController extends Controller
               else
               {
                      return response()->json(["hours"=>0]);
+              }
               }
           }
 
@@ -277,12 +312,16 @@ class EnseignantController extends Controller
        public function UpdateMyEmail( UpdateEnseignantRequest $request )
 
           {
+
+                 if (Gate::allows('ens_check_id', $id)) {
                  $id=auth()->user()->enseignant->id;
                  $enseignant=Enseignant::where('id',$id)->first();  
                  $enseignant->email_perso=$request['email_perso'];   
                  $enseignant->save();  
                  return $this->succes("","email updated successfully");
           }
+          return $this->error('','ACCES INTERDIT ',403);
+       }
 
                 
 //
